Track scroller until mouse button is released.
This uses one of Apple's methodologies for tracking mouse drag events. It
prevents from additional events to be processed by the Application until the
mouse button has been released.

On Mac OS X 10.7, there was a strange visual effect to the mouse cursor while
it was tracking a drag. Once the mouse moved near the borders of the window the
window sizer cursor would appear. In the normal NSScrollView, this does not
happen. Since this patch blocks additional events from being processed, then
the mouse cursor never gets a chance to change to the window sizer cursor until
the mouse button is released.