Pop animation group from stack before committing.
Fixes inadvertent calls to the wrong animationDidStop selector by popping the
last animation group that is being committed off the animation group stack
before committing the animation. This ensure that any implicit animations added
during the commit are not added to the last animation group, instead they are
added to the previous animation group on the stack.